+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+## 2026-09-04
+
+### Features
+
+* count a customer as active if they've been seen in the last year ([272991e](https://adl.github.com/adl-developer/kinkane-backend/commit/272991e840b180a017553a1096fda60d674a6488)) — [details](changelog/2026-09-04-active-customer-sign-in.md)
+
+
 ## 2026-09-03
 
 ### Features